+++ tests/unit/components/settings/EditCustomer.spec.js
@@ -86,9 +86,9 @@ describe("EditCustomer", () => {
 
       expect(id_error_message.text()).toStrictEqual(invalid_text);
       expect(pass_error_message.text()).toStrictEqual(invalid_text);
-      expect(user_account_id_error_message.text()).toStrictEqual(invalid_text);
+      expect(user_account_id_error_message.text()).toStrictEqual("");
     });
-    test.each(["passkey-id", "alphanumeric-id", "user-account-id"])(
+    test.each(["passkey-id", "alphanumeric-id"])(
       "When EditCustomer has invalid credentials, Then both ID and passkey will mount with invalid text and will both become become valid with any change to %s",
       async (selector_id_suffix) => {
         const id_error_message = wrapper.find("#input-widget-feedback-alphanumeric-id");
@@ -98,7 +98,7 @@ describe("EditCustomer", () => {
 
         expect(id_error_message.text()).toStrictEqual(invalid_text);
         expect(pass_error_message.text()).toStrictEqual(invalid_text);
-        expect(user_account_id_error_message.text()).toStrictEqual(invalid_text);
+        expect(user_account_id_error_message.text()).toStrictEqual("");
 
         const target_input_field = wrapper.find("#input-widget-field-" + selector_id_suffix);
         await target_input_field.setValue("new entry");
@@ -128,14 +128,6 @@ describe("EditCustomer", () => {
       ["Cat * lab", "passkey", "passkey-id", "validate_customer_account_input"],
       ["Valid", "passkey", "passkey-id", "validate_customer_account_input"],
       ["Cat lab", "passkey", "passkey-id", "validate_customer_account_input"],
-      [
-        "Experiment anemia alpha cells -1",
-        "user_account_id",
-        "user-account-id",
-        "validate_customer_account_input",
-      ],
-      ["C", "user_account_id", "user-account-id", "validate_customer_account_input"],
-      ["", "user_account_id", "user-account-id", "validate_customer_account_input"],
     ])(
       "When the text %s (%s) is entered into the field found with the selector ID %s, Then the correct text validation function (%s) is called and the error message from the validation function is rendered below the input in the DOM",
       async (entry, text_id, selector_id_suffix, text_validation_type) => {
@@ -177,7 +169,6 @@ describe("EditCustomer", () => {
     test.each([
       ["alphanumeric-id", "This field is required"],
       ["passkey-id", "This field is required"],
-      ["user-account-id", "This field is required"],
     ])(
       "Given some nonsense value in the input field with the DOM Id suffix %s, When the input field is updated to be a blank value, Then the error message below the text in the DOM matches what the business logic dictates (%s)",
       async (selector_id_suffix, expected_message) => {